requirements.txt
Django==2.0
djangorestframework==3.8.2
mpmath==1.0.0
pillow==5.2.0
psycopg2==2.6.2
pyserial==3.4
pytz==2018.5
selenium==3.141.0
sympy==1.1.1
urllib3==1.24.1
virtualenv==16.0.0
